requirements.txt
langchain>=0.1.0,<0.3.0
langchain-core>=0.1.0,<0.3.0
langchain-community>=0.0.10,<0.3.0
langchain-groq>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests
python-dotenv
streamlit>=1.20.0
